requirements.txt
fastapi==0.65.2
orjson
uvicorn[standard]
gunicorn
tortoise-orm[aiomysql]
tortoise-orm[accel]
python-jose[cryptography]
python-multipart